Simplified write permission check
The old method for checking if a file is writtable will fail if the
file has a write lock, yet a file can be deleted (on unix) even if it
has a write lock. Now we only check that the file is not read only.
There are cases in which some read only file is not writtable (for
example when the user has no permissions). In that case, the
installation will fail later on and will be rolled back.