Fix database update issue
When doing a full refresh of the addin db, the whole db is deleted,
so data about domains assigned to folders is deleted. In this case,
when generating the new data, the domain of a folder (and the
addins it contains) may change. When reseting the cached data,
we now update the domains of the already loaded addins, so that
they don't try to read the description using the old domain id.