Fix add-in update issue
When an add-in is updated when it has already been loaded, it is
registered for uninstallation and then disabled. Since we now disable
add-ins without taking account the version, the add-in that we are
installing (the update) also gets disabled. The solution is to
disable the uninstalled add-in using the add-in version, only for
that specific use case.