Fix issue when upgrading core add-ins
The UpgradeCodeAddin test was broken since the logic for enablig/disabling
add-ins has changed. Disabling an add-in now disables all versions of the
add-in. When replacing a core add-in, the add-in can't be uninstalled, so it
has to be disabled, but it has to be done using an overload of the DisableAddin
method that disables that specific version of the add-in, not all versions.