Type Extension Points didn't work correctly, when using CustomExtensionAttribute. Using CustomExtensionAttribute on several TypeExtensionNodes didn't assign extension nodes to the correct extension points. Therefore, calling AddInEngine.GetExtensionNodes(typeof(IAnyInterface)) also returned extension nodes of another interface.
Equivalent to ExtensionAttribute, CustomExtensionAttribute with property Type now.
CustomAttribute with a list of nodes and properties (@IAssemblyReflector.cs)
Reflector.cs/DefaultAssemblyReflector.cs now returns a list of properties and a list of nodes. Properties are needed to get value of property Type.
AddinScanner now checks property Type of CustomExtensionAttribute, otherwise (equivalent to ExtensionAttribute) the base types are analyzed and used as path.