Cap length of FileDatabase filenames at 130.
On Linux filesystems, the maximum filename length is 255 bytes. However, when using home directory encryption (as opposed to full-disk encryption), user-visible filenames are backed by encrypted filenames which are much longer. As a result, filenames longer than about 143 characters cannot be stored on encrypted systems.

A limit of 130 gives room for a fairly lengthy file extension.

This limit, unlike MAX_PATH on Windows, is per-file not for the full path name.