Allow custom addin localizer to be defined with attribute
An addin can now replace the localizer used when translating strings
by specifying the type in the AddinLocalizerGettext attribute.

[assembly:AddinLocalizerGettext(Type="MyNamespace.MyAddinLocalizer")]

Previously the type would always default to Gettext and would result
in the Mono.Addins.Localization.GettextLocalizer used. It was already
possible to use a custom addin localizer if the addin info is defined
in the addin.xml, but not possible using attributes to define the
addin info.