Optimize the cecil reflector
The reflector now takes into account that some assemblies only have to
be loaded if they reference Mono.Addins. For example, when looking for
a subclass of Mono.Addins.ExtensionNode, only assemblies that have a
Mono.Addins reference can define such class.
This optimization can be applied in several cases and reduces the
number of assemblies that are loaded.

An UnloadAssembly method has also been added, and it is used to unload
potential add-in assemblies when there is an error while scanning.