Adds support for generating add-in data cache files
Added a new command and api that can be used to generate a file next to
each add-in file that contains the result of the scan. The next time the
add-in needs to be scanned, the scanner will load the data from the data
file instead of scanning again the add-in.