Avoid loading the addin if it has no localizer
When translating strings defined in an addin.xml file the addin
and its dependencies will be loaded. If conditions or extension
objects used this will result in addin assemblies being loaded.
If the addin does not have a localizer then the AddinEngine's
default localizer will be used and the addin does not need to
be loaded just to translate the string.