Support using HttpClient when downloading files
By default the HttpClient will be used when downloading files such
as the addin repository xml and any custom images for the addin
manager dialog. Also a custom HttpClient factory can be supplied
instead of the default HttpClient creation. The WebRequestHelper
is still supported if a custom web request handler is being used
instead.