Delay add-in assembly load by lazy creating objects
Delay the lookup of condition and localizer types until the objects
are actually going to be used. Also avoid loading types for
type based extension points. Types were loaded to map them to
corresponding extension paths, but the mapping can be done
using the type name instead of the actual type instance.